feat: add reload at logo and change file name
