+++ newsstand/src/views/header/header.css
@@ -13,6 +13,8 @@
 .logo-container {
     display: flex;
     gap: 8px;
+
+    cursor: pointer;
 }
 
 .logo-label {
