+++ newsstand/src/views/header/header.js
@@ -1,23 +1,12 @@
-export const Header = () => {
+import { getDate } from "../../utils/utils.js";
+
+export const HeaderContainer = (handleReload) => {
   let element = document.createElement("div");
   element.className = "header-container";
   
-  function getDate() {
-    const currentDate = new Date();
-
-    const year = currentDate.getFullYear();
-    const month = String(currentDate.getMonth() + 1).padStart(2, "0");
-    const day = String(currentDate.getDate()).padStart(2, "0");
-
-    const daysOfWeek = ["일요일", "월요일", "화요일", "수요일", "목요일", "금요일", "토요일"];
-    const dayOfWeek = daysOfWeek[currentDate.getDay()]; // 0부터 일요일
-
-    return `${year}. ${month}. ${day}. ${dayOfWeek}`;
-  }
-
   function render() {
     const html = `
-        <div class="logo-container">
+        <div class="logo-container" id="logo-container">
             <img src="../../assets/icons/newspaper.svg" alt="newspaper icon">
             <span class="logo-label">뉴스스탠드</span>
         </div>
@@ -27,11 +16,20 @@ export const Header = () => {
     element.innerHTML = html;
   }
 
+  function addEventListeners() {
+    const logoContainer = element.querySelector("#logo-container");
+
+    if (logoContainer) {
+      logoContainer.addEventListener("click", handleReload);
+    }
+  }
+
   render();
+  addEventListeners()
 
   return {
     element,
   };
 };
 
-export default Header;
+export default HeaderContainer;
